Intcode is now 64-bit, in prep for day9

+++ src/day5.rs
@@ -2,15 +2,15 @@ use aoc_runner_derive::{aoc, aoc_generator};
 use crate::intcode::{Program, IOOperation};
 
 #[aoc_generator(day5)]
-fn day5_gen(input: &str) -> Vec<i32> {
+fn day5_gen(input: &str) -> Vec<i64> {
     input
         .split(',')
-        .flat_map(|num_str| num_str.parse::<i32>())
+        .flat_map(|num_str| num_str.parse::<i64>())
         .collect()
 }
 
 #[aoc(day5, part1)]
-pub(crate) fn part1(input: &[i32]) -> i32 {
+pub(crate) fn part1(input: &[i64]) -> i64 {
     let mut output = -6969;
     Program::from(input).run(|io_op| {
         match io_op {
@@ -27,7 +27,7 @@ pub(crate) fn part1(input: &[i32]) -> i32 {
 }
 
 #[aoc(day5, part2)]
-pub(crate) fn part2(input: &[i32]) -> i32 {
+pub(crate) fn part2(input: &[i64]) -> i64 {
     let mut output = -6969;
     Program::from(input).run(|io_op| {
         match io_op {
